+++ app/Http/Controllers/Api/EmpleadoController.php
@@ -7,6 +7,8 @@ use App\Http\Resources\EmpleadoCollection; // Importar la colección EmpleadoCol
 use App\Http\Requests\StoreEmpleadosRequest; // Importar la request StoreEmpleadosRequest
 use App\Http\Requests\UpdateEmpleadosRequest; // Importar la request UpdateEmpleadosRequest
 use Symfony\Component\HttpFoundation\Response; // Importar la clase Response para los códigos de estado HTTP
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests; // Importar el trait AuthorizesRequests para la autorización de politicas
+
 
 
 use App\Http\Controllers\Controller;
@@ -16,6 +18,7 @@ use App\Models\Empleado; // Importar el modelo Empleado
 
 class EmpleadoController extends Controller
 {
+    use AuthorizesRequests;
      // Muestra todos los empleados
     public function index(){
         // return Categoria::all(); // Devuelve todos los clientes
@@ -31,7 +34,8 @@ class EmpleadoController extends Controller
 
     // Almacena un nuevo empleado 
     public function store(StoreEmpleadosRequest $request){  // Usar la request StoreEmpeadosRequest para validar los datos
-        $empleado = Empleado::create($request->all());  // Crear un nuevo empleado con los datos validados    
+        //$empleado = Empleado::create($request->all());  // Crear un nuevo empleado con los datos validados    
+        $empleado = $request->user()->empleados()->create($request->all());  // Crear un nuevo empleado asociada al usuario autenticado
        
         // Devolver el empleado creado como recurso API con código de estado 201 (creado) 
         return response()->json(new EmpleadoResource($empleado), Response::HTTP_CREATED); 
@@ -39,6 +43,7 @@ class EmpleadoController extends Controller
 
      // Actualiza un empleado existente
     public function update(UpdateEmpleadosRequest $request, Empleado $empleado){  // Usar la request UpdateEmpleadosRequest para validar los datos
+        $this->authorize('update', $empleado);  // Autorizar la acción usando la política EmpleadoPolicy
         $empleado->update($request->all());  // Actualizar el empleado con los datos validados
 
         // Devolver el empleado actualizado como recurso API con código de estado 200 (OK)
@@ -47,6 +52,7 @@ class EmpleadoController extends Controller
 
      // Elimina un empleado existente
     public function destroy(Empleado $empleado){  // Inyectar el empleado a eliminar
+        $this->authorize('delete', $empleado);  // Autorizar la acción usando la política EmpleadoPolicy
         $empleado->delete();  // Eliminar el empleado
 
         // Devolver una respuesta vacía con código de estado 204 (No Content)
